+++ ariadne_codegen/client_generators/result_types.py
@@ -1,8 +1,9 @@
 import ast
-from typing import Dict, List, Optional, Tuple, cast
+from typing import Dict, List, Optional, Set, Tuple, cast
 
 from graphql import (
     DirectiveNode,
+    ExecutableDefinitionNode,
     FieldNode,
     FragmentDefinitionNode,
     FragmentSpreadNode,
@@ -30,6 +31,7 @@ from ..codegen import (
     generate_import_from,
     generate_method_call,
     generate_module,
+    generate_pass,
 )
 from ..exceptions import NotSupported, ParsingError
 from ..plugins.manager import PluginManager
@@ -58,8 +60,9 @@ class ResultTypesGenerator:
     def __init__(
         self,
         schema: GraphQLSchema,
-        operation_definition: OperationDefinitionNode,
+        operation_definition: ExecutableDefinitionNode,
         enums_module_name: str,
+        fragments_module_name: Optional[str] = None,
         fragments_definitions: Optional[Dict[str, FragmentDefinitionNode]] = None,
         base_model_import: Optional[ast.ImportFrom] = None,
         convert_to_snake_case: bool = True,
@@ -70,8 +73,10 @@ class ResultTypesGenerator:
         self.operation_definition = operation_definition
         if not self.operation_definition.name:
             raise NotSupported("Operations without name are not supported.")
+        self._operation_name = self.operation_definition.name.value
 
         self.enums_module_name = enums_module_name
+        self.fragments_module_name = fragments_module_name
         self.fragments_definitions = (
             fragments_definitions if fragments_definitions else {}
         )
@@ -86,44 +91,45 @@ class ResultTypesGenerator:
             or generate_import_from([BASE_MODEL_CLASS_NAME], PYDANTIC_MODULE),
         ]
         self._public_names: List[str] = []
-        self._class_defs: List[ast.ClassDef] = []
         self._used_enums: List[str] = []
         self._used_scalars: List[str] = []
         self._used_fragments_names: set[str] = set()
 
         self._class_defs = self._parse_type_definition(
-            class_name=str_to_pascal_case(self.operation_definition.name.value),
-            type_name=self._get_operation_type_name(
-                self.operation_definition.operation
-            ),
+            class_name=str_to_pascal_case(self._operation_name),
+            type_name=self._get_operation_type_name(self.operation_definition),
             selection_set=self.operation_definition.selection_set,
         )
+        self._add_enums_scalars_fragments_imports()
 
-    def _get_operation_type_name(self, operation_type: OperationType) -> str:
-        if operation_type == OperationType.QUERY and self.schema.query_type:
+    def _get_operation_type_name(self, definition: ExecutableDefinitionNode) -> str:
+        if isinstance(definition, FragmentDefinitionNode):
+            return definition.type_condition.name.value
+
+        if (
+            isinstance(definition, OperationDefinitionNode)
+            and definition.operation == OperationType.QUERY
+            and self.schema.query_type
+        ):
             return self.schema.query_type.name
 
-        if operation_type == OperationType.MUTATION and self.schema.mutation_type:
+        if (
+            isinstance(definition, OperationDefinitionNode)
+            and definition.operation == OperationType.MUTATION
+            and self.schema.mutation_type
+        ):
             return self.schema.mutation_type.name
 
         if (
-            operation_type == OperationType.SUBSCRIPTION
+            isinstance(definition, OperationDefinitionNode)
+            and definition.operation == OperationType.SUBSCRIPTION
             and self.schema.subscription_type
         ):
             return self.schema.subscription_type.name
 
-        raise NotSupported(f"Not supported operation type: {operation_type}")
+        raise NotSupported(f"Not supported operation type: {definition}")
 
     def generate(self) -> ast.Module:
-        if self._used_enums:
-            self._imports.append(
-                generate_import_from(self._used_enums, self.enums_module_name, 1)
-            )
-        if self._used_scalars:
-            for scalar_name in self._used_scalars:
-                scalar_data = self.custom_scalars[scalar_name]
-                self._imports.extend(generate_scalar_imports(scalar_data))
-
         update_forward_refs_calls = [
             generate_expr(
                 generate_method_call(class_def.name, UPDATE_FORWARD_REFS_METHOD)
@@ -143,10 +149,16 @@ class ResultTypesGenerator:
             )
         return module
 
+    def get_imports(self) -> List[ast.ImportFrom]:
+        return self._imports
+
+    def get_classes(self) -> List[ast.ClassDef]:
+        return self._class_defs
+
     def get_operation_as_str(self) -> str:
         operation_str = print_ast(self.operation_definition)
         if self._used_fragments_names:
-            for used_fragment in sorted(self._used_fragments_names):
+            for used_fragment in sorted(self._get_all_related_fragments()):
                 operation_str += "\n\n" + print_ast(
                     self.fragments_definitions[used_fragment]
                 )
@@ -160,6 +172,9 @@ class ResultTypesGenerator:
     def get_generated_public_names(self) -> List[str]:
         return self._public_names
 
+    def get_used_fragments(self) -> List[str]:
+        return list(self._used_fragments_names)
+
     def _parse_type_definition(
         self,
         class_name: str,
@@ -168,17 +183,13 @@ class ResultTypesGenerator:
         add_typename: bool = False,
         extra_bases: Optional[List[str]] = None,
     ) -> List[ast.ClassDef]:
-        class_bases = [BASE_MODEL_CLASS_NAME]
-        if extra_bases:
-            class_bases.extend(extra_bases)
-        class_def = generate_class_def(class_name, class_bases)
-
-        if class_def.name in self._public_names:
+        if class_name in self._public_names:
             return []
-        self._public_names.append(class_def.name)
+        self._public_names.append(class_name)
 
-        extra_classes = []
-        resolved_selection_set = self._resolve_selection_set(selection_set, type_name)
+        resolved_selection_set, fragments = self._resolve_selection_set(
+            selection_set, type_name
+        )
         if add_typename:
             (
                 resolved_selection_set,
@@ -186,6 +197,16 @@ class ResultTypesGenerator:
             ) = self._add_typename_field_to_selections(
                 resolved_selection_set, selection_set
             )
+
+        if fragments:
+            class_bases = [str_to_pascal_case(f) for f in sorted(fragments)]
+        else:
+            class_bases = [BASE_MODEL_CLASS_NAME]
+        if extra_bases:
+            class_bases.extend(extra_bases)
+        class_def = generate_class_def(class_name, class_bases)
+
+        extra_classes = []
         for lineno, field in enumerate(
             resolved_selection_set,
             start=1,
@@ -234,29 +255,31 @@ class ResultTypesGenerator:
                     operation_definition=self.operation_definition,
                     selection_set=selection_set,
                 )
+
+        if not class_def.body:
+            class_def.body.append(generate_pass())
+
         return [class_def] + extra_classes
 
     def _resolve_selection_set(
         self, selection_set: SelectionSetNode, root_type: str = ""
-    ) -> List[FieldNode]:
+    ) -> Tuple[List[FieldNode], Set[str]]:
         fields = []
+        fragments = set()
         for selection in selection_set.selections:
             if isinstance(selection, FieldNode):
                 fields.append(selection)
             elif isinstance(selection, FragmentSpreadNode):
-                self._used_fragments_names.add(selection.name.value)
-                fields.extend(
-                    self._resolve_selection_set(
-                        self.fragments_definitions[selection.name.value].selection_set,
-                        root_type,
-                    )
-                )
+                fragments.add(selection.name.value)
             elif isinstance(selection, InlineFragmentNode):
                 if selection.type_condition.name.value == root_type:
-                    fields.extend(
-                        self._resolve_selection_set(selection.selection_set, root_type)
+                    sub_fields, sub_fragments = self._resolve_selection_set(
+                        selection.selection_set, root_type
                     )
-        return fields
+                    fields.extend(sub_fields)
+                    fragments = fragments.union(sub_fragments)
+        self._used_fragments_names = self._used_fragments_names.union(set(fragments))
+        return fields, fragments
 
     def _add_typename_field_to_selections(
         self, resolved_fields: List[FieldNode], selection_set: SelectionSetNode
@@ -369,3 +392,50 @@ class ResultTypesGenerator:
         for field_type_name in field_types_names:
             if field_type_name.type_name in self.custom_scalars:
                 self._used_scalars.append(field_type_name.type_name)
+
+    def _add_enums_scalars_fragments_imports(self):
+        if self._used_enums:
+            self._imports.append(
+                generate_import_from(self._used_enums, self.enums_module_name, 1)
+            )
+        if self._used_scalars:
+            for scalar_name in self._used_scalars:
+                scalar_data = self.custom_scalars[scalar_name]
+                self._imports.extend(generate_scalar_imports(scalar_data))
+
+        if (
+            isinstance(self.operation_definition, OperationDefinitionNode)
+            and self._used_fragments_names
+            and self.fragments_module_name
+        ):
+            self._imports.append(
+                generate_import_from(
+                    [str_to_pascal_case(f) for f in self._used_fragments_names],
+                    self.fragments_module_name,
+                    1,
+                )
+            )
+
+    def _get_all_related_fragments(self) -> Set[str]:
+        fragments_names: Set[str] = self._used_fragments_names.copy()
+        for fragment_name in self._used_fragments_names:
+            fragment_def = self.fragments_definitions[fragment_name]
+            fragments_names = fragments_names.union(
+                self._get_fragments_names(fragment_def.selection_set)
+            )
+        return fragments_names
+
+    def _get_fragments_names(self, selection_set: SelectionSetNode) -> Set[str]:
+        names: Set[str] = set()
+        for node in selection_set.selections:
+            if isinstance(node, FragmentSpreadNode):
+                name = node.name.value
+                names.add(name)
+                names = names.union(
+                    self._get_fragments_names(
+                        self.fragments_definitions[name].selection_set
+                    )
+                )
+            elif isinstance(node, FieldNode) and node.selection_set:
+                names = names.union(self._get_fragments_names(node.selection_set))
+        return names