+++ ariadne_codegen/plugins/base.py
@@ -2,13 +2,14 @@ import ast
 from typing import Dict, Optional, Tuple, Union
 
 from graphql import (
+    ExecutableDefinitionNode,
     FieldNode,
+    FragmentDefinitionNode,
     GraphQLEnumType,
     GraphQLInputField,
     GraphQLInputObjectType,
     GraphQLSchema,
     Node,
-    OperationDefinitionNode,
     SelectionSetNode,
     VariableDefinitionNode,
 )
@@ -88,13 +89,13 @@ class Plugin:
 
     # pylint: disable=unused-argument
     def generate_result_types_module(
-        self, module: ast.Module, operation_definition: OperationDefinitionNode
+        self, module: ast.Module, operation_definition: ExecutableDefinitionNode
     ) -> ast.Module:
         return module
 
     # pylint: disable=unused-argument
     def generate_operation_str(
-        self, operation_str: str, operation_definition: OperationDefinitionNode
+        self, operation_str: str, operation_definition: ExecutableDefinitionNode
     ) -> str:
         return operation_str
 
@@ -102,7 +103,7 @@ class Plugin:
     def generate_result_class(
         self,
         class_def: ast.ClassDef,
-        operation_definition: OperationDefinitionNode,
+        operation_definition: ExecutableDefinitionNode,
         selection_set: SelectionSetNode,
     ) -> ast.ClassDef:
         return class_def
@@ -111,7 +112,7 @@ class Plugin:
     def generate_result_field(
         self,
         field_implementation: ast.AnnAssign,
-        operation_definition: OperationDefinitionNode,
+        operation_definition: ExecutableDefinitionNode,
         field: FieldNode,
     ) -> ast.AnnAssign:
         return field_implementation
@@ -149,3 +150,11 @@ class Plugin:
     # pylint: disable=unused-argument
     def process_name(self, name: str, node: Optional[Node] = None) -> str:
         return name
+
+    # pylint: disable=unused-argument
+    def generate_fragments_module(
+        self,
+        module: ast.Module,
+        fragments_definitions: Dict[str, FragmentDefinitionNode],
+    ) -> ast.Module:
+        return module