+++ tests/conftest.py
@@ -10,9 +10,11 @@ def mocked_plugin_manager(mocker):
     manager.generate_client_code.side_effect = no_effect_method
     manager.generate_enums_code.side_effect = no_effect_method
     manager.generate_inputs_code.side_effect = no_effect_method
+    manager.generate_result_class.side_effect = no_effect_method
     manager.generate_result_types_code.side_effect = no_effect_method
     manager.copy_code.side_effect = no_effect_method
     manager.generate_scalars_code.side_effect = no_effect_method
     manager.generate_init_code.side_effect = no_effect_method
     manager.process_name.side_effect = no_effect_method
+    manager.generate_fragments_module.side_effect = no_effect_method
     return manager