+++ tests/main/clients/multiple_fragments/expected_client/__init__.py
@@ -1,0 +1,48 @@
+from .async_base_client import AsyncBaseClient
+from .base_model import BaseModel
+from .client import Client
+from .example_query1 import ExampleQuery1, ExampleQuery1ExampleQuery
+from .example_query2 import ExampleQuery2, ExampleQuery2ExampleQuery
+from .example_query3 import ExampleQuery3, ExampleQuery3ExampleQuery
+from .exceptions import (
+    GraphQLClientError,
+    GraphQLClientGraphQLError,
+    GraphQLClientGraphQLMultiError,
+    GraphQLClientHttpError,
+    GraphQlClientInvalidResponseError,
+)
+from .fragments import (
+    CompleteA,
+    CompleteAFieldB,
+    FullA,
+    FullAFieldB,
+    FullB,
+    MinimalA,
+    MinimalAFieldB,
+    MinimalB,
+)
+
+__all__ = [
+    "AsyncBaseClient",
+    "BaseModel",
+    "Client",
+    "CompleteA",
+    "CompleteAFieldB",
+    "ExampleQuery1",
+    "ExampleQuery1ExampleQuery",
+    "ExampleQuery2",
+    "ExampleQuery2ExampleQuery",
+    "ExampleQuery3",
+    "ExampleQuery3ExampleQuery",
+    "FullA",
+    "FullAFieldB",
+    "FullB",
+    "GraphQLClientError",
+    "GraphQLClientGraphQLError",
+    "GraphQLClientGraphQLMultiError",
+    "GraphQLClientHttpError",
+    "GraphQlClientInvalidResponseError",
+    "MinimalA",
+    "MinimalAFieldB",
+    "MinimalB",
+]